clean up
added a new notebook which explains a lot of the project and how to run
the code. removed a lot of excess code that was generated during
development. Now will start writing tests for the code and maybe
another simpler explanation notebook which will describe how to use the
diffusion solvers